Split GDDLRobtopLevelsLayer
